+++ NEWS.md
@@ -1,5 +1,7 @@
 # Version 1.0.9000.9028
-* Renamed the `maxOverdoseProb` slot to `max_overdose_prob` in `NextBestNCRM` class.
+* Renamed several slots in `NextBestNCRM` and `NextBestDualEndpoint` classes.
+* Changed the `scale` character slot into `target_relative` flag slot in
+  `NextBestDualEndpoint` class.
 * Renamed the argument of `derive` function from `mtdSamples` to `mtd_samples` in
   `NextBestMTD` class.
 * Allowed for `from_prior` flag - argument to `modelspecs` function at `GeneralModel`